decoder finalized

+++ VAssembler/vasm_logging.hpp
@@ -7,4 +7,4 @@
 #include "vasm_flags.hpp"
 
 extern void print_err(const std::string& str);
-extern void print_info(const std::string& str, size_t required_info_level = 1);
+extern void print_info(const std::string& str, int required_info_level = 1);
